pyproject.toml
[tool.poetry]
name = "kitsune"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35.0"
numpy = "^1.26.4"
scipy = "^1.13.1"
pyshark = "^0.6"
pymongo = "^4.7.3"
matplotlib = "^3.9.0"
cython = "^3.0.10"
pandas = "^2.2.2"
dpkt = "^1.9.8"
pcapyplus = "^1.0.1"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.0"

click = "^8.1.7"
tqdm = "^4.66.4"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kitsune = "kitsune.cli:main"